+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/server/security/DatabaseWhitelistAccessControlHandler.java
@@ -16,9 +16,7 @@
 package com.hotels.bdp.waggledance.server.security;
 
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -27,7 +25,7 @@ import com.hotels.bdp.waggledance.api.model.PrimaryMetaStore;
 
 public class DatabaseWhitelistAccessControlHandler implements AccessControlHandler {
 
-  private final Set<Pattern> databaseWhiteList = new HashSet<>();
+  private final List<Pattern> databaseWhiteList = new ArrayList<>();
   private final FederationService federationService;
   private PrimaryMetaStore primaryMetaStore;
   private final boolean hasCreatePermission;
@@ -58,9 +56,6 @@ public class DatabaseWhitelistAccessControlHandler implements AccessControlHandl
       return true;
     }
     databaseName = trimToLowerCase(databaseName);
-    if (databaseWhiteList.contains(databaseName)) {
-      return true;
-    }
 
     for (Pattern whiteListEntry : databaseWhiteList) {
       Matcher matcher = whiteListEntry.matcher(databaseName);