+++ waggle-dance-core/src/test/java/com/hotels/bdp/waggledance/server/security/DatabaseWhitelistAccessControlHandlerTest.java
@@ -46,7 +46,7 @@ public class DatabaseWhitelistAccessControlHandlerTest {
   @Mock
   private FederationService federationService;
   private DatabaseWhitelistAccessControlHandler handler;
-  private final List<String> whitelist = newArrayList("writabledb", "regexdb.*");
+  private final List<String> whitelist = newArrayList("writabledb", "userdb.*");
 
   @Before
   public void setUp() {
@@ -63,9 +63,9 @@ public class DatabaseWhitelistAccessControlHandlerTest {
 
   @Test
   public void hasRegexGrantedWritePermission() throws Exception {
-    assertTrue(handler.hasWritePermission("regexDB1"));
-    assertTrue(handler.hasWritePermission("regexdb2"));
-    assertFalse(handler.hasWritePermission("regex"));
+    assertTrue(handler.hasWritePermission("userDB1"));
+    assertTrue(handler.hasWritePermission("userdb2"));
+    assertFalse(handler.hasWritePermission("user"));
   }
 
   @Test
@@ -82,7 +82,7 @@ public class DatabaseWhitelistAccessControlHandlerTest {
     verify(federationService).update(eq(primaryMetaStore), captor.capture());
     PrimaryMetaStore updatedMetastore = captor.getValue();
     assertThat(updatedMetastore.getWritableDatabaseWhiteList().size(), is(3));
-    assertThat(updatedMetastore.getWritableDatabaseWhiteList(), contains("writabledb", "regexdb.*", "newdb"));
+    assertThat(updatedMetastore.getWritableDatabaseWhiteList(), contains("writabledb", "userdb.*", "newdb"));
   }
 
   @Test
@@ -92,7 +92,7 @@ public class DatabaseWhitelistAccessControlHandlerTest {
     verify(federationService).update(eq(primaryMetaStore), captor.capture());
     PrimaryMetaStore updatedMetastore = captor.getValue();
     assertThat(updatedMetastore.getWritableDatabaseWhiteList().size(), is(2));
-    assertThat(updatedMetastore.getWritableDatabaseWhiteList(), contains("writabledb", "regexdb.*"));
+    assertThat(updatedMetastore.getWritableDatabaseWhiteList(), contains("writabledb", "userdb.*"));
   }
 
 }